backend/requirements.txt
numpy<2
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
geopandas==0.14.2
shapely==2.0.2
fiona==1.9.5
pyproj==3.6.1
httpx==0.26.0
pydantic>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
psycopg2-binary>=2.9.0
